Tests\Integration: updated to support new signatures KP-8034

+++ src/Tests/WB.Tests.Integration/InterviewTests/LinkedQuestions/when_answering_linked_single_option_question_which_links_to_real_question_and_which_is_roster_title_for_2_rosters_and_roster_level_is_1.cs
@@ -32,9 +32,9 @@ namespace WB.Tests.Integration.InterviewTests.LinkedQuestions
             var linkedOption1Vector = new decimal[] { 0 };
             linkedOption2Vector = new decimal[] { 1 };
             var linkedOption3Vector = new decimal[] { 2 };
-            var linkedOption1Answer = (decimal) 6.5;
-            var linkedOption2Answer = (decimal) 3.54;
-            var linkedOption3Answer = (decimal) -11.2;
+            var linkedOption1Answer = 6.5;
+            var linkedOption2Answer = 3.54;
+            var linkedOption3Answer = -11.2;
             linkedOption2TextInvariantCulture = "3.54";
 
             var triggerQuestionId = Guid.NewGuid();
